extract and split integration tests
